Refactor picker.iconpicker
