+++ app/src/main/java/com/myoxidae/moneez/picker/iconpicker/IconListAdapter.kt
@@ -1,4 +1,4 @@
-package com.myoxidae.moneez.picker
+package com.myoxidae.moneez.picker.iconpicker
 
 import android.content.Context
 import android.view.LayoutInflater
